// packages/commercial-lytenyte-pro/src/components/external/floating-ui.d.ts
declare type Coords = { [key in Axis]: number };
declare type Dimensions = { [key in Length]: number };
declare interface ElementRects {
  reference: Rect;
  floating: Rect;
}
declare type Length = "width" | "height";
declare type Padding = number | Prettify$1<Partial<SideObject>>;
declare type Placement = Prettify$1<Side | AlignedPlacement>;
declare type Prettify$1<T> = { [K in keyof T]: T[K] } & {};
declare type Rect = Prettify$1<Coords & Dimensions>;
declare type Side = "top" | "right" | "bottom" | "left";
declare type SideObject = { [key in Side]: number };
declare type Strategy = "absolute" | "fixed";
// ...
declare interface ComputePositionConfig$1 {
  /**
   * Object to interface with the current platform.
   */
  platform: Platform$1;
  /**
   * Where to place the floating element relative to the reference element.
   */
  placement?: Placement;
  /**
   * The strategy to use when positioning the floating element.
   */
  strategy?: Strategy;
  /**
   * Array of middleware objects to modify the positioning or provide data for
   * rendering.
   */
  middleware?: Array<Middleware$1 | null | undefined | false>;
}
declare interface ComputePositionReturn extends Coords {
  /**
   * The final chosen placement of the floating element.
   */
  placement: Placement;
  /**
   * The strategy used to position the floating element.
   */
  strategy: Strategy;
  /**
   * Object containing data returned from all middleware, keyed by their name.
   */
  middlewareData: MiddlewareData;
}
// ...
declare type Middleware$1 = {
  name: string;
  options?: any;
  fn: (state: MiddlewareState$1) => Promisable$1<MiddlewareReturn>;
};
declare interface MiddlewareData {
  [key: string]: any;
  arrow?: Partial<Coords> & {
    centerOffset: number;
    alignmentOffset?: number;
  };
  autoPlacement?: {
    index?: number;
    overflows: Array<{
      placement: Placement;
      overflows: Array<number>;
    }>;
  };
  flip?: {
    index?: number;
    overflows: Array<{
      placement: Placement;
      overflows: Array<number>;
    }>;
  };
  hide?: {
    referenceHidden?: boolean;
    escaped?: boolean;
    referenceHiddenOffsets?: SideObject;
    escapedOffsets?: SideObject;
  };
  offset?: Coords & {
    placement: Placement;
  };
  shift?: Coords & {
    enabled: { [key in Axis]: boolean };
  };
}
declare interface MiddlewareReturn extends Partial<Coords> {
  data?: {
    [key: string]: any;
  };
  reset?:
    | boolean
    | {
        placement?: Placement;
        rects?: boolean | ElementRects;
      };
}
// ...
declare interface ShiftOptions$1 extends DetectOverflowOptions$1 {
  /**
   * The axis that runs along the alignment of the floating element. Determines
   * whether overflow along this axis is checked to perform shifting.
   * @default true
   */
  mainAxis?: boolean;
  /**
   * The axis that runs along the side of the floating element. Determines
   * whether overflow along this axis is checked to perform shifting.
   * @default false
   */
  crossAxis?: boolean;
  /**
   * Accepts a function that limits the shifting done in order to prevent
   * detachment.
   */
  limiter?: {
    fn: (state: MiddlewareState$1) => Coords;
    options?: any;
  };
}
// ...
declare const computePosition: (
  reference: ReferenceElement,
  floating: FloatingElement,
  options?: Partial<ComputePositionConfig>,
) => Promise<ComputePositionReturn>;
declare type ComputePositionConfig = Prettify<
  Omit<ComputePositionConfig$1, "middleware" | "platform"> & {
    /**
     * Array of middleware objects to modify the positioning or provide data for
     * rendering.
     */
    middleware?: Array<Middleware | null | undefined | false>;
    /**
     * Custom or extended platform object.
     */
    platform?: Platform;
  }
>;
declare type Derivable<T> = (state: MiddlewareState) => T;
// ...
declare type DetectOverflowOptions = Prettify<
  Omit<DetectOverflowOptions$1, "boundary"> & {
    boundary?: Boundary;
  }
>;
// ...
declare type FloatingElement = HTMLElement;
// ...
declare type Middleware = Prettify<
  Omit<Middleware$1, "fn"> & {
    fn(state: MiddlewareState): Promisable<MiddlewareReturn>;
  }
>;
/**
 * @deprecated use `MiddlewareState` instead.
 */
declare type MiddlewareArguments = MiddlewareState;
declare type MiddlewareState = Prettify<
  Omit<MiddlewareState$1, "elements"> & {
    elements: Elements;
  }
>;
declare interface NodeScroll {
  scrollLeft: number;
  scrollTop: number;
}
/**
 * Modifies the placement by translating the floating element along the
 * specified axes.
 * A number (shorthand for `mainAxis` or distance), or an axes configuration
 * object may be passed.
 * @see https://floating-ui.com/docs/offset
 */
declare const offset: (options?: OffsetOptions) => Middleware;
declare type OffsetOptions = OffsetValue | Derivable<OffsetValue>;
declare type OffsetValue =
  | number
  | {
      /**
       * The axis that runs along the side of the floating element. Represents
       * the distance (gutter or margin) between the reference and floating
       * element.
       * @default 0
       */
      mainAxis?: number;
      /**
       * The axis that runs along the alignment of the floating element.
       * Represents the skidding between the reference and floating element.
       * @default 0
       */
      crossAxis?: number;
      /**
       * The same axis as `crossAxis` but applies only to aligned placements
       * and inverts the `end` alignment. When set to a number, it overrides the
       * `crossAxis` value.
       *
       * A positive number will move the floating element in the direction of
       * the opposite edge to the one that is aligned, while a negative number
       * the reverse.
       * @default null
       */
      alignmentAxis?: number | null;
    };
declare interface Platform {
  getElementRects: (args: {
    reference: ReferenceElement;
    floating: FloatingElement;
    strategy: Strategy;
  }) => Promisable<ElementRects>;
  getClippingRect: (args: {
    element: Element;
    boundary: Boundary;
    rootBoundary: RootBoundary;
    strategy: Strategy;
  }) => Promisable<Rect>;
  getDimensions: (element: Element) => Promisable<Dimensions>;
  convertOffsetParentRelativeRectToViewportRelativeRect: (args: {
    elements?: Elements;
    rect: Rect;
    offsetParent: Element;
    strategy: Strategy;
  }) => Promisable<Rect>;
  getOffsetParent: (
    element: Element,
    polyfill?: (element: HTMLElement) => Element | null,
  ) => Promisable<Element | Window>;
  isElement: (value: unknown) => Promisable<boolean>;
  getDocumentElement: (element: Element) => Promisable<HTMLElement>;
  getClientRects: (element: Element) => Promisable<Array<ClientRectObject>>;
  isRTL: (element: Element) => Promisable<boolean>;
  getScale: (element: HTMLElement) => Promisable<{
    x: number;
    y: number;
  }>;
}
declare const platform: Platform;
declare type Prettify<T> = { [K in keyof T]: T[K] } & {};
declare type Promisable<T> = T | Promise<T>;
declare type ReferenceElement = Element | VirtualElement;
/**
 * Optimizes the visibility of the floating element by shifting it in order to
 * keep it in view when it will overflow the clipping boundary.
 * @see https://floating-ui.com/docs/shift
 */
declare const shift: (options?: ShiftOptions | Derivable<ShiftOptions>) => Middleware;
declare type ShiftOptions = Prettify<Omit<ShiftOptions$1, "boundary"> & DetectOverflowOptions>;
// ...
declare interface VirtualElement {
  getBoundingClientRect(): ClientRectObject;
  getClientRects?(): Array<ClientRectObject> | DOMRectList;
  contextElement?: Element;
}